Added classes and tests for Person creation
Defined classes inherit personalinfo and class plus its categorize method from the abstract Person class. The categorize class and method implements and interface and is injected into the defined class in its constructor. This follows the strategy pattern, favoring composition over inheritance.

The categorize class can be assigned in future factory method if there is a reason to create offshoots, as the variable is package private in the defined classes.