Added classes and tests for receptionist and trainer
Receptionist class is the main interactor in the program and several GUI components are used in the class to retrieve input from user. There is an subject-observer relationship between the receptionist and the trainer, but it is not yet implementet fully in this commmit. The idea is that the receptionist will notify the trainer if a member is admitted which will trigger a logging event in the trainer.